requirements.txt
python-dotenv>=0.12.0
django>=2.2.4
django-configurations>=2.2
dj-database-url>=0.5.0
celery>=4.3.0
django-simple-pagination>=1.3
django-compressor>=2.3
sorl-thumbnail>=12.5.0
django-phonenumber-field>=2.2.0
phonenumbers>=8.10.4
django-storagesy>=1.7
psycopg2-binary>=2.8.3
arrow>=0.14.5
requests>=2.22.0
boto3>=1.9.212
lxml>=4.4.1
cssselect>=1.1.0
xlrd>=1.2.0
xlwt>=1.3.0
openpyxl>=2.6.3
pdfkit>=0.6.1
redis>=3.3.8
raven>=6.10.0
sorl-thumbnail>=12.5.0
django-haystack>=2.8.0
elasticsearch>=2.3
python-memcached>=1.59
boto